refactor: 일정 추가 컴포넌트 네이밍 변경
